+++ config/commands/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/plugin-commands-config
 
+## 1100.0.5
+
+### Patch Changes
+
+- Updated dependencies [0fbcf74]
+  - @pnpm/config.reader@1101.1.2
+  - @pnpm/workspace.workspace-manifest-writer@1100.0.3
+
 ## 1100.0.4
 
 ### Patch Changes